Move address to separate namespace
