+++ source/Energinet.DataHub.MeteringPoints.Domain/Addresses/Rules/MunicipalityCodeRuleError.cs
@@ -14,32 +14,20 @@
 
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Rules
+namespace Energinet.DataHub.MeteringPoints.Domain.Addresses.Rules
 {
-    public class CityNameLengthRule : IBusinessRule
+    public class MunicipalityCodeRuleError : ValidationError
     {
-        private const int MaxLength = 25;
-
-        public CityNameLengthRule(string? city)
+        public MunicipalityCodeRuleError()
         {
-            Validate(city);
+            MunicipalityCode = default;
         }
 
-        public bool IsBroken { get; private set; }
-
-        public ValidationError ValidationError { get; private set; } = new CityNameLengthRuleError();
-
-        private void Validate(string? city)
+        public MunicipalityCodeRuleError(int municipalityCode)
         {
-            if (string.IsNullOrWhiteSpace(city))
-            {
-                return;
-            }
-
-            if (city.Length <= MaxLength) return;
-
-            IsBroken = true;
-            ValidationError = new CityNameLengthRuleError(city, MaxLength);
+            MunicipalityCode = municipalityCode;
         }
+
+        public int MunicipalityCode { get; }
     }
 }
