+++ source/Energinet.DataHub.MeteringPoints.Domain/MeteringPoints/MarketMeteringPoint.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System;
+using Energinet.DataHub.MeteringPoints.Domain.Addresses;
 using Energinet.DataHub.MeteringPoints.Domain.GridAreas;
 using NodaTime;
