+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/EDI/Errors/Converters/Address/CityNameLengthRuleErrorConverter.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System;
+using Energinet.DataHub.MeteringPoints.Domain.Addresses.Rules;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Rules;
 
 namespace Energinet.DataHub.MeteringPoints.Infrastructure.EDI.Errors.Converters.Address
