+++ source/Energinet.DataHub.MeteringPoints.Tests/Domain/MeteringPoints/MarketMeteringPointTests.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System.Linq;
+using Energinet.DataHub.MeteringPoints.Domain.Addresses;
 using Energinet.DataHub.MeteringPoints.Domain.GridAreas;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
