+++ source/Energinet.DataHub.MeteringPoints.Tests/Domain/MeteringPoints/MeteringPointTests.cs
@@ -14,6 +14,7 @@
 
 using System.Linq;
 using Azure.Core;
+using Energinet.DataHub.MeteringPoints.Domain.Addresses;
 using Energinet.DataHub.MeteringPoints.Domain.GridAreas;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
